Added some arithmatic instructions

+++ src/Main.java
@@ -6,14 +6,14 @@ import static value.StringValue.Value;
 import java.util.Arrays;
 
 import runtime.Engine;
-import value.IntValue;
+import value.DoubleValue;
 
 public class Main {
 	public static void main(String[] args) {
 		new Engine().run(Arrays.asList(
 			Load(Value("print")),
 			Push(Value("Hello World!")),
-			Push(IntValue.Value(1)),
+			Push(DoubleValue.Value(1)),
 			Call()
 		));
 	}
